refactor: creating a Fixed Priority Scheduler
